feat/#1: board 컴포넌트 분리
